Separate SpritePackerBase from ImagePacker.
